backend/requirements/development.txt
-r ./base.txt

django-extensions==3.1.5

pytest==8.0.2
pytest-cov==3.0.0
pytest-django==4.5.1
pytest-mock==3.12.0

ipdb==0.13.11
Faker==24.9.0
factory_boy==3.2.1
freezegun==1.2.2

Sphinx==5.3.0
django-rest-framework
django-cors-headers